implement autorization using gates

+++ app/Http/Controllers/QuestionController.php
@@ -63,9 +63,12 @@ class QuestionController extends Controller
      * @param  \App\Question  $question
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Question $question)
     {   
-        $question = Question::findOrFail($id);
+        // $question = Question::findOrFail($id);
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied !');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -77,7 +80,11 @@ class QuestionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(AskQuestionRequest $request, Question $question)
-    {
+    {   
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied !');
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
     }
@@ -89,7 +96,11 @@ class QuestionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {   
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied !');
+        }
+
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
     }
